disfluencies handling
-When the CONT from MS is borrowed to use for PTB tokens, the length of
the disfluency also changes.  To account for this, the PTB disfluency is
kept the same BUT instances might be repeated (for instance when
'sixty-two' is read as 'sixty' 'two', the a 'C' label is repeated).
This will capture the disfluency label in relative good faith, but could
introduce problems with targeting a specific 'IP' label, for instance.