disfluencies updates and bug fixing
 - Two outputs will be produced where the dtype (e.g. 'ptb') is printed
as part of the output file (data.py)

 -  The error types are updated for bug fixing, MIX is an option and the
logic to choose a summary is improved (errinfo.py)

 - A bug which caused dict values in ErrInfo objects to print incorrectly
is fixed (errinfo.py)

 - Disfluencies info is being recorded, and the logic to get previous
disfluencies is functioning (main.py, errinfo.py)

 - Instead of, i.e. leaving out 'DEL' type errors while scraping PTB
info, now 'DEL' errors would be left in IF they are part of a
'PTB'-style error (e.g. 'SUB' + 'DEL').  This means the next token is
ALWAYS the same token for 'PTB' and 'MS'.